Updated to Please Sir Ben W.
Added the one Else i missed

Erased boyd output 2 code since object class will already make sure the PWM outputs are in the right range

double checked that code will turn us the right way- it will

updated class to make it save some memory -still cant compile the damn thing- HELP

+++ motorclass.cpp
@@ -46,7 +46,7 @@ void motor::direction(bool fwdRev);{
 void speed(int desiredSpeed){
   // funciton sets the desired PWM value
 
-  int cappedSpeed; // Value within arduino PWM range;
+  uint8_t cappedSpeed; // Value within arduino PWM range;
 
   if(desiredSpeed > 255){
     // check for overflow
